Moved ObjectChildResource to relations group
